Triangulation of Polygons with multiple Holes possible now
The Demo Project was updated to show the possibility to handle more than
one hole in the Triangulator.
The Split Polygon function for the Triangulator was rewritten to be able
to split up the Polygon into all subpolygons at once (no recursive
calls).
TODO: Need to enhance performance! It it now a lot slower than before