Bug Fix for issue #1340 - BackMaterial added
If a ScreenSpaceVisual3D Element is mirrord at the YZ Plane then the triangle indecies are calculated in reverse order. This means that we see the back side of the mesh. Unfortunately the BackMaterial was not set. So the Visuals become invisible. The fix just set the Material and the BackMaterial to the same  value.